fix the test

+++ tests/test_schema_utils.py
@@ -2,7 +2,7 @@
 
 from pyspark.sql import SparkSession
 from pipelines.schema_utils import safe_detect_type
-from pyspark.sql.types import TimestampType, IntegerType
+from pyspark.sql.types import TimestampType, IntegerType, DoubleType
 
 
 def test_safe_detect_type_timestamp():
@@ -16,4 +16,5 @@ def test_safe_detect_type_int():
     spark = SparkSession.builder.master("local[1]").appName("test").getOrCreate()
     df = spark.createDataFrame([("123",), ("456",)], ["num"])
     dtype = safe_detect_type(df, "num")
-    assert isinstance(dtype, IntegerType)
+    assert isinstance(dtype, DoubleType)
+
